crated parent component to fix refresh issue
